// src/app/models/enuns/Status.ts
export enum Status  {
    NEW = "",
    ATIVO = "Ativo",
    INATIVO = "Inativo",
    RECESSO ="Recesso", 
    PENDENTE = "Ativo",
    ABERTO = "Aberto",
    APROVADO = "Aprovado",
    CANCELADO = "Analise",
    ESTORNO = "Estornado",
    PROCESSANDO = "Processando",
    PAY = "Aguardando pagamento",
    CONCLUIDO  = "Concluido",
    ANALISE  = "Aguardando confirmação",
    STOCK_ERROR = "Erro de estoque"
  }
 